microbit-targets: corrected spaces in directory error
Previously the CMake file would pass an unescaped string to the
command line. If this string had a directory containing spaces
the linker would process this the argument as a list as opposed
to a directory string.

This commit introduces escaped quotes passed directory passed to the
command line indicating that the argument is a directory string.