yotta-deprecation: update dependency of mbed-gcc to use git
As the yotta registry is deprecated (as per #6) we need to move away
from using it.

All of these targets have a dependency on the mbed-gcc target, which
was being fetched by the registry (defined through 'inherits'). This
patch changes the reference to explicitly use GitHub as the source
for the dependency.

This uses an HTTPS URL, which yotta will fetch with the GitHub API,
while this can cause some issues with rate limiting, it also means
a user doesn't need to setup SSH at all, which is required for
git@ URLS